requirements.txt
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
fastapi
uvicorn[standard]
python-dotenv
pypdf